Created main_menu scene and MenuScript
Brought TextMeshPro into the project (that's why large number of changed files). Created a main_menu scene with a canvas that has the title, credits, copyright, and start and exit buttons. Created MenuScript that is to be used with the game manager. Used debug to unsure the buttons are linked to their methods in the MenuScript.